Test with 7 modules (restore previous functionality)
Well, I only have 3 so far, but I can see that everything seems to be
fine as the buffer is rotated and displays correctly on my three
modules.
This should conclude the modifications to get the same functionality as
before, except now with a counter driven interrupt and no ugly active
waiting or delays.