Cleanup/fixup pio assembly code
Actually the bitrate was not quite correct as there was an extra
instruction not accounted for in the computation.
Actually, maybe there is not much meaning in having the delays in the
first place? Perhaps the clock has some influence on how quick we can
get data into the fifo? Anyhow, it works in the current state. I'll
leave it for now.