Use timer instead of active wait for row illumination duration
Had to use the SM stall on empty RX fifo instead of the previous wait
for IRQ flag cleared as the handler now is split into two functions.
If the interrupt flags are not cleared in the first one, it will be
called again just after returning from it. If we however clear the
interrupt flags in the first handler (before the timer) it makes no
sense to have the SM to wait for them.

This change now also allows for other tasks, especially the UART
transfers, to be processed during the row illumination wait (explicitly
verified with very long illumination duration).

In the current state however the illumination duration actually still is
the PIO serial data transmission time + the timer duration. There might
be a way to circumvent this. Perhaps I'll try to implement it at some
point.