Add minimal code required for TinyUSB CDC
TinyUSB may be used to transmit data via USB port on the pico board. We
cannot use interrupts as with UART but this is fine as we may run all
TinyUSB on the second core of the MCU.
The UART option is retained.
usb_descriptors.c and tusb_config.h are for the most part copied over
from tinyusb official cdc_msc example. MSC is not used as of now but we
may use it at some point if sensible. Those sections specific to MCUs
other than the pi pico ones (rp2040, rp2350) have been removed.
Actually, tinyusb MCU should be set to rp2040 even for the rp2350.