Add script for processing frame for multi module row displays
Has to do some snaked unwrapping of the image to get it into the correct
format. For now, this is not done on the microcontroller. I guess it may
be possible, but if not really required maybe not the highest priority
for now.
This would be required if we want the current method of scrolling to
still work. Perhaps however it is better to just use a larger frame
buffer? Then we are not limited to scrolling but can display whatever
animation (which fits the buffer).